# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ciph3r Product API",
        "description": "API RESTful para la gestión de productos, divisas y precios de productos.",
        "contact": {
            "email": "tu_email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Ciph3r API Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener lista de productos",
                "description": "Retorna una lista paginada de productos, con opción de filtros y ordenación. Incluye la divisa base y todos los precios asociados a cada producto.",
                "operationId": "06ef3dce57df563227f6458dafe547dc",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar productos por nombre (búsqueda parcial)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Filtrar productos por descripción (búsqueda parcial)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "currency_id",
                        "in": "query",
                        "description": "Filtrar productos por el ID de su divisa base",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Campo para ordenar la lista (ej. 'name', 'price')",
                        "schema": {
                            "type": "string",
                            "default": "name"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Dirección de la ordenación ('asc' o 'desc')",
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear un nuevo producto",
                "description": "Crea un nuevo producto y lo retorna. Los campos `price`, `tax_cost` y `manufacturing_cost` deben ser numéricos. `currency_id` debe existir en la tabla de divisas.",
                "operationId": "fdbc5bd92f75ba29dd69b783d402ba91",
                "requestBody": {
                    "description": "Datos del producto a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener producto por ID",
                "description": "Retorna un producto específico por su ID. Incluye la divisa base y todos los precios asociados.",
                "operationId": "a08dea864fe850c94992df85a744cf59",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto obtenido exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Actualizar producto",
                "description": "Actualiza un producto existente por su ID y lo retorna. Permite actualizar parcialmente el producto. Los campos `price`, `tax_cost` y `manufacturing_cost` deben ser numéricos. `currency_id` debe existir en la tabla de divisas.",
                "operationId": "f6f7fa856f17792f2ba4ef3c020fe576",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del producto a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name has already been taken."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Eliminar producto",
                "description": "Elimina un producto por su ID. También elimina todos los precios asociados a este producto.",
                "operationId": "dc4b0c092a8e42fbb802f8bd4b8288bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Producto eliminado exitosamente (No Content)."
                    },
                    "404": {
                        "description": "Producto no encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Currency": {
                "title": "Currency",
                "description": "Modelo de una divisa",
                "properties": {
                    "id": {
                        "description": "Identificador único de la divisa",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Nombre de la divisa (ej. US Dollar)",
                        "type": "string"
                    },
                    "symbol": {
                        "description": "Símbolo de la divisa (ej. USD)",
                        "type": "string"
                    },
                    "exchange_rate": {
                        "description": "Tasa de cambio de la divisa respecto a una base (ej. 1.0 para USD)",
                        "type": "number",
                        "format": "float"
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha y hora de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "title": "Product",
                "description": "Modelo de un producto",
                "properties": {
                    "id": {
                        "description": "Identificador único del producto",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Nombre del producto",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descripción del producto",
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "description": "Precio del producto en la divisa base",
                        "type": "number",
                        "format": "float"
                    },
                    "currency_id": {
                        "description": "Identificador de la divisa base",
                        "type": "integer"
                    },
                    "tax_cost": {
                        "description": "Costo de impuestos del producto",
                        "type": "number",
                        "format": "float"
                    },
                    "manufacturing_cost": {
                        "description": "Costo de fabricación del producto",
                        "type": "number",
                        "format": "float"
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha y hora de última actualización",
                        "type": "string",
                        "format": "date-time"
                    },
                    "currency": {
                        "$ref": "#/components/schemas/Currency"
                    },
                    "prices": {
                        "description": "Lista de precios del producto en diferentes divisas",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductPrice"
                        }
                    }
                },
                "type": "object"
            },
...
            "StoreProductRequest": {
                "title": "Store Product Request",
                "description": "Datos para crear un nuevo producto",
                "required": [
                    "name",
                    "price",
                    "currency_id",
                    "tax_cost",
                    "manufacturing_cost"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre único del producto",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Smartphone X Pro"
                    },
                    "description": {
                        "description": "Descripción detallada del producto",
                        "type": "string",
                        "example": "Un smartphone de última generación con cámara de 108MP.",
                        "nullable": true
                    },
                    "price": {
                        "description": "Precio base del producto",
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 799.990000000000009094947017729282379150390625
                    },
                    "currency_id": {
                        "description": "ID de la divisa base del producto. Debe existir en la tabla 'currencies'.",
                        "type": "integer",
                        "example": 1
                    },
                    "tax_cost": {
                        "description": "Costo de impuestos aplicables al producto",
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 120.5
                    },
                    "manufacturing_cost": {
                        "description": "Costo de fabricación del producto",
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 450
                    }
                },
                "type": "object"
            },
            "UpdateProductRequest": {
                "title": "Update Product Request",
                "description": "Datos para actualizar un producto existente (parcialmente).",
                "properties": {
                    "name": {
                        "description": "Nombre único del producto",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Smartphone X Pro (Updated)"
                    },
                    "description": {
                        "description": "Descripción detallada del producto",
                        "type": "string",
                        "example": "Una descripción actualizada del smartphone X Pro.",
                        "nullable": true
                    },
                    "price": {
                        "description": "Precio base del producto",
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 829.990000000000009094947017729282379150390625
                    },
                    "currency_id": {
                        "description": "ID de la divisa base del producto. Debe existir en la tabla 'currencies'.",
                        "type": "integer",
                        "example": 2
                    },
                    "tax_cost": {
                        "description": "Costo de impuestos aplicables al producto",
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 125
                    },
                    "manufacturing_cost": {
                        "description": "Costo de fabricación del producto",
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 460
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Autenticación con Laravel Sanctum. Genera un token usando /api/auth/login y envíalo en el encabezado 'Authorization: Bearer {token}'.",
                "name": "Sanctum Token based Auth",
                "in": "header",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}